Update CloseOrderWindow.java
Since values 10, 50, 100 and 200 can be pressed multiple times, duplicates can occur and based on how many taps are pressed, the amount of receipts printed will be based on the number of taps. To avoid that, the buttons are only set to be able to be press once and only enabled again if another value is tapped. If the user however chooses to pay Exact amount (This is before the Calculator section), all of the potential misclicks are also disabled to ensure no multiple taps/duplicates will occur.