Delimiter and Tax Fix (Indonesia)
Delimiters for Rp did not occur (Reported by Fika Coffee) on tax and also the entire receipt does not have delimiters (1000 will be displayed instead of 1,000). This fix intends to remove the .00 at the taxes and adds delimiters to the entire receipt. This will only apply if the currency is set to Rp (Rupiah). The setting revert back to normal if any other currency than Rp is used.