Updated Pax Promotion Logic
The previous logic will cause issues where the prices are not updated accordingly. A better solution would be to see if the promotion exists on the order as promotion pax can only exist once on an order and never more than one because promotionweight will determine the pax promotion to be used.