Discount Fix
Now when discount via percentage or value is done on the first split item, users can no longer discount the order on the second split. This means user needs to void the entire bill if they made a mistake on the splitted discount. Previously, when discounting on split orders, the result will not tally in the end session report. This is due to the orderDetails set to the discount of the entire order discount, not having it's own individual discounts.