Discount Update
When you change the quantity of the item, the discount of the item does not update accordingly. This is fixed by applying the exact logic used to set the discount of the item. Another fix done was the addition of another item after a discounted item split is closed. Prevoiusly, when discounted item split is closed and a new item is added, the discount is wrong and the X report was incorrect. This is fixed by deducting the total discount left and making sure everything is calculated right after the deduction.